@@ -1,5 +1,5 @@
 import { Injectable } from "@angular/core";
-import { Http } from "@angular/http";
+import { HttpClient } from "@angular/common/http";
 
 import { MEAT_API } from "app/app.api";
 import { Observable } from "rxjs";
@@ -7,38 +7,30 @@ import "rxjs/add/operator/map";
 import "rxjs/add/operator/catch";
 
 import { Restaurant } from "./restaurant/restaurant.model";
-import { ErrorHandler } from "app/app.error-handler";
+//import { ErrorHandler } from "app/app.error-handler";
 import { MenuItem } from "app/restaurant-detail/menu-item/menu-item.model";
 
 @Injectable()
 export class RestaurantsService {
-  constructor(private http: Http) {}
+  constructor(private http: HttpClient) {}
 
   restaurants(): Observable<Restaurant[]> {
     return this.http
-      .get(`${MEAT_API}/restaurants`)
-      .map((response) => response.json())
-      .catch(ErrorHandler.handleError);
+      .get<Restaurant[]>(`${MEAT_API}/restaurants`)
   }
 
   restaurantById(id: string): Observable<Restaurant> {
     return this.http
-      .get(`${MEAT_API}/restaurants/${id}`)
-      .map((response) => response.json())
-      .catch(ErrorHandler.handleError);
+      .get<Restaurant>(`${MEAT_API}/restaurants/${id}`)
   }
 
   reviewsOfRestaurant(id: string): Observable<any> {
     return this.http
       .get(`${MEAT_API}/restaurants/${id}/reviews`)
-      .map((response) => response.json())
-      .catch(ErrorHandler.handleError);
   }
 
   menuOfRestaurant(id: string): Observable<MenuItem[]> {
     return this.http
-      .get(`${MEAT_API}/restaurants/${id}/menu`)
-      .map((response) => response.json())
-      .catch(ErrorHandler.handleError);
+      .get<MenuItem[]>(`${MEAT_API}/restaurants/${id}/menu`)
   }
 }
